// frontend/src/app/types/api.d.ts
export type components = {
    schemas: {
        BaseResponse: {
            status: components["schemas"]["Status"];
        };
        Client: {
            active: boolean;
            /** Format: date-time */
            creationDate: string;
            id: number;
            identityNumber: string;
            name: string;
            phone: string;
        };
// ...
        Reservation: {
            active: boolean;
            charge: number;
            /** Format: date-time */
            checkInDate: string;
            /** Format: date-time */
            checkOutDate: string;
            clientId: number;
            /** Format: date-time */
            creationDate: string;
            id: number;
            roomId: number;
            /** @enum {string} */
            status: "Pending" | "Confirmed" | "CheckedIn" | "CheckedOut" | "Cancelled";
        };
// ...
    };
// ...
};